Allow a type property on a resource type.
