+++ lib/raml/node/abstract_resource_circular.rb
@@ -1,0 +1,3 @@
+class Raml::AbstractResource < Raml::PropertiesNode
+  child_of :type, [ Raml::ResourceType, Raml::ResourceTypeReference ]
+end
